Refactor Loadable HOC
